Environment Path value will be appended for the level(User ,Machine and process )
Validates User value with value available in "Path" Environment variable for given Level(User,Machine)and append the value to it if path value is not available.
Expands User value with value available in "Path" Environment variable for given Level(User,Machine)and append the value to it if path value is not available.
Creates new Path variable if it is not available for given  Level(User,Machine) and append the value to it